Move reusable sticky note components to lib-components/employee/notes
